requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
telethon==1.34.0
python-dotenv
